src/test/java/com/sg/superperson2/test/service/SuperpersonServiceTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.sg.superperson2.test.service;

import javax.inject.Inject;

import org.junit.Test;
import static org.junit.Assert.assertTrue;
import org.junit.runner.RunWith;

import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.transaction.annotation.Transactional;

import com.sg.superperson2.exception.*;
import com.sg.superperson2.model.Superperson;
import com.sg.superperson2.service.SuperpersonService;

/**
 *
 * @author main
 */
@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(locations = {"classpath:test-applicationContext.xml"})
@Transactional
public class SuperpersonServiceTest {
    
    @Inject
    SuperpersonService supService;
    
    @Test
    @Transactional
    public void testInvalidSuperperson()
	    throws DuplicateObjectException {
	Superperson sup = new Superperson();
	
	boolean thrown = false;
	try {
	    supService.addSuperperson(sup);
	}
	catch (InvalidObjectException ex) {
	    thrown = true;
	}
	
	assertTrue(thrown);
    }
}
